requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.0.0b3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.3
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==2.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
slack==0.0.2
slackclient==2.9.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.6.3
